Crypto Avatars Selection boxes & Arrows for cryptoAvatars
Fixed the styles on the boxes and the buttons to match the design. Changed the refsMap part to fetch the new refs when the avatars are loaded and can be use to move the arrow on selection. Also added Id's to the lists so when you move to another list the arrow on the previous one disappears.