Main scene freezing/crashing on selecting certain avatars.
The problem was in the character menu avatar canvas. Where the code targets the head to preview in the canvas. And as some VRM files or most of them don't have the head mesh bone named exactly "Head", the neckBone was not found and that crushed the main scene as well. SO replaced === "Head" with .Includes("Head"). Worked like a charm.