change to chracterhups for remote players
When chatting with another player, the diorama in the character hups window for the remote player is buggy. The character is out of frame in the character hups for remote player. to set up side camera in diorama.js, position, quaternion and scale is obtained from target.matrixworld.decompose which does not seem to be updated for remote player. Hence calculating/composing it here with obtainted position and quaternion data